// src/types/constants.ts
export enum COLUMN_TYPE {
  Text = 'text',
  Number = 'number',
  Amount = 'amount',
  Date = 'date',
  DateTime = 'datetime',
  Boolean = 'boolean',
  Percentage = 'percentage',
  TextProfile = 'text_profile',
  Link = 'link',
  Status = 'status',
  StatusSwitch = 'status_switch',
  Custom = 'custom',
  Switch = 'switch',
}

export enum Role {
  Admin = 'admin',
  User = 'user',
  Member = 'member',
}